requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.50.0
django-extensions==3.2.0
django-filter==22.1
django-phonenumber-field==6.3.0
django-simple-history==3.1.1
djangorestframework==3.13.1
drf-spectacular==0.23.1
drf-spectacular-sidecar==2022.8.1
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
jsonschema==4.14.0
Markdown==3.4.1
oauthlib==3.2.0
phonenumbers==8.12.54
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.8.1
